Base for ConcurrentCouchbaseCheck.
This works, but could use enhancement.  The idea here is that with
nothing other than a JVM (at least for the built dist) one can
verify their Couchbase Server works with this client.

There's lots of room for improvement.  Don't count failures.  Run
threads for a stated elapsed time.  etc.